% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acf_by_block.R
\name{acf_by_block}
\alias{acf_by_block}
\title{Block-wise autocorrelation function}
\usage{
acf_by_block(x, blocks, lag_max = NULL, make_plot = TRUE, ...)
}
\arguments{
\item{x}{Numeric vector}

\item{blocks}{Vector of indices of blocks, of same length as x}

\item{lag_max}{Maximum lag at which to calculate the ACF. Defaults to the
length of the shortest block.}

\item{make_plot}{Logical: should the ACF be plotted? Defaults to TRUE.}
}
\value{
Vector of ACF at each time lag
}
\description{
This original version of this function was written by Stacy DeRuiter.
It computes the ACF of a vector passed as input, ignoring time lags that
overlap blocks (i.e., time series).
}
