% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtag.R
\name{add_RL}
\alias{add_RL}
\title{Add received level columns to DTag data frame}
\usage{
add_RL(data, RL)
}
\arguments{
\item{data}{Data frame, at least with columns 'tagID' and 'time'}

\item{RL}{Data frame of received level, with with columns 'tagID',
'time', and 'received_level'}
}
\value{
Input data frame with additional column for received level
}
\description{
Add received level columns to DTag data frame
}
