% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_by_ID.R
\name{diff_by_ID}
\alias{diff_by_ID}
\title{First-order difference for grouped data}
\usage{
diff_by_ID(x, ID)
}
\arguments{
\item{x}{Vector of values}

\item{ID}{Vector of group IDs, of same length as x}
}
\value{
Vector of same length as x of first-order differences of x
(and NAs for last value of each group ID)
}
\description{
First-order difference for grouped data
}
