% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtag.R
\name{expo_data}
\alias{expo_data}
\title{Add exposure columns to DTag data frame}
\usage{
expo_data(data, expo)
}
\arguments{
\item{data}{Data frame, at least with columns 'tagID', 'ID' (dive ID),
and 'time'}

\item{expo}{Data frame with one row for each exposure event, and one column
each for 'tagID', 'start' and 'end'}
}
\value{
Input data frame with additional columns for
\itemize{
\item{expo}{either 'before', 'during' or 'after' (an exposure event)}
\item{expo_dive}{'yes' for exposure dives, 'no' for other dives}
\item{expo_time}{time since start of exposure (0 before)}
}
}
\description{
Add exposure columns to DTag data frame
}
