% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/langevin.R
\name{grad_log_pi}
\alias{grad_log_pi}
\title{Gradient of log(pi(x)) in Langevin model}
\usage{
grad_log_pi(pt, beta, cov_list)
}
\arguments{
\item{pt}{2d point}

\item{beta}{Vector of selection coefficients}

\item{cov_list}{List of rasters}
}
\value{
2d gradient
}
\description{
Gradient of log(pi(x)) in Langevin model
}
