% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_gradient.R
\name{grad_raster}
\alias{grad_raster}
\title{Gradient of raster using bilinear interpolation}
\usage{
grad_raster(rast, pts)
}
\arguments{
\item{rast}{Raster object}

\item{pts}{Matrix with two columns (one for each coordinate)}
}
\value{
Matrix with same dimensions as pts, where the first column
gives the gradient along the first dimension, and the second column
along the second dimension.
}
\description{
Gradient of raster using bilinear interpolation
}
\details{
The gradient has a closed form solution if bilinear interpolation
is used, as described in Appendix C of Michelot (2019, PhD thesis).
https://etheses.whiterose.ac.uk/23688/
}
