% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_gradient.R
\name{grad_terra}
\alias{grad_terra}
\title{Gradient of terra raster using bilinear interpolation}
\usage{
grad_terra(rast = NULL, pts, xgrid = NULL, ygrid = NULL, zmat = NULL)
}
\arguments{
\item{rast}{SpatRast object}

\item{pts}{Matrix with two columns (one for each coordinate)}

\item{xgrid}{Grid of x values (if rast not provided)}

\item{ygrid}{Grid of y values (if rast not provided)}

\item{zmat}{Matrix of raster values (if rast not provided)}
}
\value{
Matrix with same dimensions as pts, where the first column
gives the gradient along the first dimension, and the second column
along the second dimension.
}
\description{
Gradient of terra raster using bilinear interpolation
}
\details{
The gradient has a closed form solution if bilinear interpolation
is used, as described in Appendix C of Michelot (2019, PhD thesis).
https://etheses.whiterose.ac.uk/23688/
}
