% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis.R
\name{LL_to_UTM}
\alias{LL_to_UTM}
\title{Transform longitude-latitude to UTM}
\usage{
LL_to_UTM(lon, lat, zone)
}
\arguments{
\item{lon}{Vector of longitudes}

\item{lat}{Vector of latitudes}

\item{zone}{UTM zone for projection}
}
\value{
Data frame with columns x (Easting) and y (Northing)
}
\description{
Transform longitude-latitude to UTM
}
