% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sde.R
\name{make_cov}
\alias{make_cov}
\title{Make covariance matrix for CTCRW simulation}
\usage{
make_cov(beta, sigma, dt)
}
\arguments{
\item{beta}{Parameter beta of the OU process}

\item{sigma}{Parameter sigma of the OU process}

\item{dt}{Time interval}
}
\description{
Make covariance matrix for CTCRW simulation
}
