% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{pal2}
\alias{pal2}
\title{moveHMM colour palette (Okabe and Ito)}
\format{
An object of class \code{character} of length 7.
}
\usage{
pal2
}
\description{
moveHMM colour palette (Okabe and Ito)
}
\keyword{datasets}
