% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rast.R
\name{plot_rast}
\alias{plot_rast}
\title{Plot raster}
\usage{
plot_rast(rast, contour = FALSE, guide = "colourbar")
}
\arguments{
\item{rast}{Raster object}

\item{contour}{Logical. If TRUE, contour lines are added. (Default: FALSE)}

\item{guide}{Argument passed to scale_fill. Defaults to "colourbar" for
continuous colour scale, and can be set to "none" to remove the scale.}
}
\value{
A ggplot object
}
\description{
Plot raster
}
