% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtag.R
\name{prep_dtags}
\alias{prep_dtags}
\title{Prepare data from several DTags}
\usage{
prep_dtags(
  files,
  start_times = NULL,
  out_freq,
  keep_intervals = NULL,
  n_cores = 1
)
}
\arguments{
\item{files}{Vector of file names}

\item{start_times}{Vector of simes for start of data in 'file' (as POSIX)}

\item{out_freq}{Required frequency of output}

\item{keep_intervals}{List where each element is a vector of two times: start
and end of time interval over which to keep the data (e.g., to exclude periods
where the tag was not on the animal). If not specified, the whole data set is
kept.}
}
\value{
Same data frame as prep_dtag
}
\description{
Prepare data from several DTags
}
