% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rast.R
\name{rast2df}
\alias{rast2df}
\title{Raster to dataframe}
\usage{
rast2df(rast)
}
\arguments{
\item{rast}{Raster object}
}
\value{
Data frame with columns x and y (coordinates), and z (raster value)
}
\description{
Raster to dataframe
}
