% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_gradient.R
\name{rast_to_xyz}
\alias{rast_to_xyz}
\title{Transform SpatRast to convenient format for grad_terra}
\usage{
rast_to_xyz(rast)
}
\arguments{
\item{rast}{SpatRast object}
}
\value{
List with elements x (grid of x values), y (grid of y values),
and z (matrix of raster values), as expected by grad_terra()
}
\description{
Transform SpatRast to convenient format for grad_terra
}
