% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regularise.R
\name{regularise}
\alias{regularise}
\title{Interpolate time series to regular time intervals}
\usage{
regularise(data, new_interval, var_names)
}
\arguments{
\item{data}{Data frame with columns for ID (time series ID),
time, and variables to interpolate}

\item{new_interval}{Time interval of interpolation, in the syntax
expected by 'seq'. For example, '5 sec', '1 min', '2 hour', etc.}

\item{var_names}{Vector of names of variables to interpolate

 @return Data frame with columns for ID, time, and interpolated
 variables}
}
\description{
This is a wrapper for the function stats::spline, which is applied
to each variable to interpolate.
}
