% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sde.R
\name{sim_BM}
\alias{sim_BM}
\title{Simulate Brownian motion}
\usage{
sim_BM(times, mu = 0, sigma = 1, z0 = 0)
}
\arguments{
\item{times}{Times of observation}

\item{mu}{Drift}

\item{sigma}{Diffusion}

\item{z0}{Initial value for simulation (defaults to 0)}
}
\value{
Data frame with columns z (simulated process) and time
}
\description{
Simulate Brownian motion
}
