% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sde.R
\name{sim_CIR}
\alias{sim_CIR}
\title{Simulate Cox-Ingersoll-Ross model}
\usage{
sim_CIR(times, mu = 1, beta = 1, sigma = 1, z0 = 1)
}
\arguments{
\item{times}{Times of observation}

\item{mu}{Mean (> 0)}

\item{beta}{Reversion (> 0)}

\item{sigma}{Diffusion (> 0)}

\item{z0}{Initial value for simulation (> 0)}
}
\value{
Data frame with columns z (simulated process) and time
}
\description{
Simulate Cox-Ingersoll-Ross model
}
