% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sde.R
\name{sim_CTCRW}
\alias{sim_CTCRW}
\title{Simulate from CTCRW process}
\usage{
sim_CTCRW(times, mu = 0, beta = 1, sigma = 1, z0 = 0)
}
\arguments{
\item{times}{Vector of times of observations}

\item{cov}{Vector of covariate values at times of observations}
}
\value{
Simulated track
}
\description{
Simulate from CTCRW process
}
