% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sde.R
\name{sim_GBM}
\alias{sim_GBM}
\title{Simulate from geometric Brownian motion}
\usage{
sim_GBM(times, mu = 1, sigma = 1, z0 = 1)
}
\arguments{
\item{times}{Times of observation}

\item{mu}{Percentage drift (> 0)}

\item{sigma}{Percentage volatility (> 0)}

\item{z0}{Initial value for simulation (> 0)}
}
\value{
Data frame with columns z (simulated process) and time
}
\description{
Simulate from geometric Brownian motion
}
