% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/langevin.R
\name{sim_langevin}
\alias{sim_langevin}
\title{Simulate from Langevin process}
\usage{
sim_langevin(times, beta, cov_list, gamma = 1, x0 = c(0, 0))
}
\arguments{
\item{times}{Vector of (numeric) times}

\item{beta}{Vector of selection coefficients}

\item{cov_list}{List of rasters}

\item{gamma}{Speed parameter}

\item{x0}{Initial point}
}
\value{
Matrix of simulated locations
}
\description{
Simulate from Langevin process
}
