% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_raster.R
\name{sim_raster}
\alias{sim_raster}
\title{Simulate random covariate field}
\usage{
sim_raster(rho, lim, res = 1)
}
\arguments{
\item{rho}{Coefficient of spatial autocorrelation}

\item{lim}{Vector of limits of the map (xmin, xmax, ymin, ymax)}

\item{res}{Grid resolution (defaults to 1)}
}
\value{
Raster object
}
\description{
Simulate random covariate field
}
