% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sde.R
\name{sim_CTCRW}
\alias{sim_CTCRW}
\title{Simulate from CTCRW process}
\usage{
sim_CTCRW(times, mu = 0, beta = 1, sigma = 1, tau = NULL, nu = NULL, z0 = 0)
}
\arguments{
\item{times}{Vector of times of observations}

\item{mu}{Mean velocity parameter}

\item{beta}{Reversion parameter of velocity}

\item{sigma}{Variance parameter of velocity}

\item{tau}{Time scale of autocorrelation parameter}

\item{nu}{Mean speed parameter}

\item{z0}{Initial value for simulation (default: 0)}
}
\value{
Simulated track
}
\description{
Simulate from CTCRW process
}
