% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{split_ts}
\alias{split_ts}
\title{Split time series data at gaps}
\usage{
split_ts(data, max_interval, min_ts_length = 0)
}
\arguments{
\item{data}{Data frame with columns 'ID' and 'time', and any other
data variables}

\item{max_interval}{Longest permissible time interval within a
time series, in seconds. The time series will be split where 
there are gaps longer than \code{max_interval}.}

\item{min_ts_length}{Minimum permissible length for a time series,
in seconds. Time series shorter than this will be excluded from
the output data frame. Defaults to 0, i.e., all time series are
kept.}
}
\value{
Data frame with same columns as input, plus a column 'ID_split'
for the ID of the time series segments after splitting.
}
\description{
Split time series data at gaps
}
